refactor(test): gameStore 統合テストを責務ごとにファイル分割する
226行1ファイルにまとまっており150行超のルールに抵触していたため、
ユースケース単位で3ファイルに分割する。
Vitest のファイル並列実行により実行時間の短縮も見込める。
